fix(cli): skip every skill whose source is not this repo, and record the ref
Two defects the two-axis review found in the sync reconcile.

The ownership rule only skipped a `local` source or a slash-bearing one, so
a skill vendored from a tool still had its body overwritten and rebranded;
only the source line survived, because upstream happened to carry the same
tool. #750 asks to "skip any skill whose source is not this repo", and a
tool name is not this repo: a fork re-vendors that skill itself, so the
body is its own too. Now anything not sourced from here is left untouched.

The ledger recorded no ref while --outdated compared against canary, but
the CLI syncs a fork from main. Every skill canary was ahead on would have
reported as drifted in a freshly synced fork, which is the false positive
#751 exists to remove. Entries now carry the ref they came from and
--outdated fetches that, falling back to main for an older entry.

Also from the review: the skills tests move to the mirrored path AGENTS.md
requires (tests/packages/cli/src/skills.test.ts, covering
packages/cli/src/skills.ts) rather than riding in convert.test.ts, sync
reuses the exported emptyReconcile instead of re-spelling it, and a test
pins the two independent digest implementations together, since a fork
ships no packages/cli to import and a one-sided edit would silently make
every skill read as untracked.
